Rename files to adhere initial modular concept, restructure a bit
Some files have not really good names to understand their real purpose.
For example: "engine" is not an engine at all, but it is just an FSM of
the whole game logic - a business logic, if you want.

Also as I see this is obvious that "shape_maps" is really a module and
it must have it's own header, describing interface of it's content, even
if it is used in just only one module - "field".